\name{toxtestD-internal}
\alias{ConfA}
\alias{TV.model}
\alias{OptNBinoA}
\alias{betaBino}
\alias{DW2s}
\alias{EstM2Sn}
\alias{loglM2S}
\alias{InterpolxA}
\alias{plan}
\title{Functions for internal usage}
\description{Functions for internal usage}

\author{Nadia Keddig and Werner Wosniok}
\keyword{internal}
